group migration started

+++ scm.py
@@ -8,32 +8,34 @@ import subprocess,os
 
 
 class Admin:
+
     def __init__(self, instance, user):
         self.instance = instance
         self.user = user
 
     def delete_all_projects(self):
-
+        
         project_keys = Instance(instance=self.instance, user=self.user).list_projects()
         undeleted_projects = []
 
         for key in project_keys:
             try:
-                Project(instance=self.instance, key=key, user=self.user).delete_project()
+                Project(instance=self.instance, user=self.user, key=key).delete_project()
             except:
                 undeleted_projects.append(key)
 
-        print(f"\nList of undeleted projects in {self.instance}:\n")
-        for key in undeleted_projects:
-            print(key)
+        if len(undeleted_projects) != 0:
+            print(f"\nList of undeleted projects in {self.instance}:\n")
+            for key in undeleted_projects:
+                print(key)
 
     def delete_all_groups(self):
-        group_list = Instance(instance=self.instance, user=self.user).list_groups()
+        group_list = Instance(instance=self.instance, user=self.user).list_all_groups()
 
         undeleted_groups = []
 
         for group_name in group_list:
-            if group_name != 'stash-users':
+            if group_name != 'stash-users' and group_name != 'DevOps':
                 try:
                     Group(instance=self.instance, user=self.user, group_name=group_name).delete_group()
                     
@@ -42,10 +44,10 @@ class Admin:
             else:
                 undeleted_groups.append(group_name)
 
-        print(f"\nList of undeleted groups in {self.instance}:\n")
-        for group_name in undeleted_groups:
-            print(group_name)
-        pass
+        if len(undeleted_groups) != 0:
+            print(f"\nList of undeleted groups in {self.instance}:\n")
+            for group_name in undeleted_groups:
+                print(group_name)
 
     def delete_users(self, user_list):
 
@@ -53,7 +55,7 @@ class Admin:
         undeleted_users = []
 
         for user in existing_users:
-            if user[0] in user_list:
+            if user[0] in user_list and user[0] != 'devops_admin':
                 try:
                     User(instance=self.instance, user=self.user, name=user[0], display_name=user[1], email=user[2]).delete_user()
                 except:
@@ -68,6 +70,14 @@ class Admin:
     def delete_all_users(self):
         pass
 
+    def list_all_projects_repos(self):
+
+        repo_list = {}
+
+        for key in Instance(instance=self.instance, user=self.user).list_projects():
+            repo_list.update({key:Project(instance=self.instance, user=self.user, key=key).list_repositories()})
+
+        return repo_list
 
 class Instance:
     def __init__(self, instance, user):
@@ -108,17 +118,18 @@ class Instance:
         return [api_call["values"][i]["key"] for i in range(int(api_call["size"]))]
 
     def project_list(self):
-        command = f"curl -k -u'{self.user}' https://{self.instance}/rest/api/1.0/projects?limit=100"
+        # command = f"curl -k -u'{self.user}' https://{self.instance}/rest/api/1.0/projects?limit=100"
+        list_projects = f"""curl -u'{self.user}' --request GET --url 'https://{self.instance}/rest/api/latest/projects?limit=200' --header 'Accept:application/json' """
 
-        api_call = self.run_post(command)
+        api_call = self.run_post(list_projects)
         api_call = self.api_to_json(api_call)
 
         projects = [api_call["values"][i] for i in range(int(api_call["size"]))]
-        return [(projects[i]["key"], projects[i]["name"]) for i in range(len(projects))]
+        return [(projects[i]["key"], projects[i]["name"], projects[i]["id"]) for i in range(len(projects))]
 
-    def list_groups(self):
+    def list_all_groups(self):
 
-            list_groups = f"""curl -u'{self.user}' --request GET --url 'https://{self.instance}/rest/api/latest/admin/groups?limit=800' --header 'Accept:application/json' """
+            list_groups = f"""curl -u'{self.user}' --request GET --url 'https://{self.instance}/rest/api/latest/admin/groups?limit=1600' --header 'Accept:application/json' """
             api_call = self.run_post(list_groups)
             api_call = self.api_to_json(api_call)
 
@@ -144,6 +155,21 @@ class Instance:
 
         return [(user["name"], user["displayName"], user["emailAddress"]) for user in users]
 
+    def create_groups(self):
+        project_list = self.list_projects()
+
+        for project_key in project_list:
+            group_names = [project_key + "_ADMIN", project_key + "_WRITE", project_key + "_READ"]
+            for group_name in group_names:
+                Group(instance=self.instance, user=self.user, group_name=group_name).create_group()
+
+    def give_group_permissions(self):
+        
+        projects = Instance(instance=self.instance, user=self.user).list_projects()
+        self.create_groups()
+
+        for project in projects:
+            Project(instance=self.instance, user=self.user, key=project).give_group_permission()
 
 class Group(Instance):
     def __init__(self, instance, user, group_name):
@@ -162,16 +188,16 @@ class Group(Instance):
 
         self.run_post(delete_group)
 
-    def mirror_groups(self, target_scm):
-        group_list_source = self.list_groups()
-        group_list_target = Instance(instance=target_scm, user=self.user).list_groups()
+    def mirror_all_groups(self, target_scm):
+        group_list_source = self.list_all_groups()
+        group_list_target = Instance(instance=target_scm, user=self.user).list_all_groups()
 
         untransferred_groups = []
 
         for group_name in group_list_source:
             if group_name not in group_list_target:
                 try:
-                    Group(instance=self.instance, user=self.user, group_name=group_name).create_group()
+                    Group(instance=target_scm, user=self.user, group_name=group_name).create_group()
                 except:
                     untransferred_groups.append(group_name)
             else:
@@ -181,6 +207,15 @@ class Group(Instance):
         for group_name in untransferred_groups:
             print(group_name)
 
+    def list_all_group_permissions(self):
+
+        projects = self.list_projects()
+        group_permissions = {}
+
+        for key in projects:
+            group_permissions.update({key:Project(instance=self.instance, user=self.user, key=key).list_group_permission()})
+
+        return group_permissions
 
 class User(Instance):
     def __init__(self, instance, user, name, display_name, email):
@@ -200,7 +235,7 @@ class User(Instance):
 
         self.run_post(delete_user)
 
-    def mirror_users(self, target_scm):
+    def mirror_all_users(self, target_scm):
 
         user_list_source = Instance(instance=self.instance, user=self.user).user_list()
         user_list_target = Instance(instance=target_scm, user=self.user).list_users()
@@ -222,7 +257,7 @@ class User(Instance):
 
         for name in untransferred_users:
             print(name)
-     
+
 
 class Project(Instance):
     def __init__(self, instance, user, key):
@@ -230,25 +265,17 @@ class Project(Instance):
         super().__init__(instance, user)
 
     def list_repositories(self):
-        command = f"curl -k -u{self.user} https://{self.instance}/rest/api/1.0/projects/{self.key}/repos?limit=100"
+
+        list_repositories = f"""curl -u'{self.user}' --request GET --url 'https://{self.instance}/rest/api/latest/projects/{self.key}/repos?limit=100' --header 'Accept:application/json' """
         try:
-            api_call = self.run_post(command)
+            api_call = self.run_post(list_repositories)
             api_call = self.api_to_json(api_call)
 
             return [api_call["values"][i]["name"] for i in range(int(api_call["size"]))]
         except:
             return []
 
-    def get_project_group_permissions(self):
-        
-        project_group_permissions = f"""curl -u'{self.user}' --request GET --url 'https://{self.instance}/rest/api/latest/projects/{self.key}/permissions/groups?limit=300' --header 'Accept:application/json' """
-
-        api_call = self.run_post(project_group_permissions)
-        api_call = self.api_to_json(api_call)
-
-        return [(api_call["values"][i]["group"]["name"],api_call["values"][i]["permission"]) for i in range(int(api_call["size"]))]
-
-    def get_project_user_permissions(self):
+    def user_permissions(self):
         
         project_user_permissions = f"""curl -u'{self.user}' --request GET --url 'https://{self.instance}/rest/api/latest/projects/{self.key}/permissions/users' --header 'Accept:application/json'"""
 
@@ -257,7 +284,7 @@ class Project(Instance):
 
         return [(api_call["values"][i]["user"]["name"],api_call["values"][i]["permission"]) for i in range(int(api_call["size"]))]
 
-    def mirror_repos(self, repo_list, target_scm):
+    def mirror_repos(self, repo_list:list, target_scm):
 
         target = Project(instance=target_scm, user=self.user, key=self.key)
 
@@ -299,7 +326,7 @@ class Project(Instance):
 
         PROJECT_URL = f'https://{self.instance}/rest/api/latest/projects'
         DICT = dict()
-        create_project = """curl --request POST -u %s --url '%s' --header 'Accept:application/json' --header 'Content-Type:application/json' --data '{"key":"%s","avatarUrl":"","avatar":"","links": %s}'"""%(self.user, PROJECT_URL, self.key ,DICT)
+        create_project = """curl --request POST -u '%s' --url '%s' --header 'Accept:application/json' --header 'Content-Type:application/json' --data '{"key":"%s","avatarUrl":"","avatar":"","links": %s}'"""%(self.user, PROJECT_URL, self.key ,DICT)
 
         try:
             self.run_post(create_project)
@@ -334,6 +361,40 @@ class Project(Instance):
         except:
             raise Exception(f"Could not delete {self.key} on {self.instance}. It may not be exists or {self.user[:self.user.index(':')]} does not have admin permission, or your project is not empty.")
 
+    def list_group_permission(self):
+
+        project_group_permissions = f"""curl -u'{self.user}' --request GET --url 'https://{self.instance}/rest/api/latest/projects/{self.key}/permissions/groups?limit=300' --header 'Accept:application/json' """
+
+        api_call = self.run_post(project_group_permissions)
+        api_call = self.api_to_json(api_call)
+
+        return [(api_call["values"][i]["group"]["name"],api_call["values"][i]["permission"]) for i in range(int(api_call["size"]))]
+
+    def give_group_permission(self):
+        groups = Instance(instance=self.instance, user=self.user).list_all_groups()
+
+        PERMISSIONS = ["PROJECT_ADMIN", "PROJECT_WRITE", "PROJECT_READ"]
+        GROUP_NAMES = [self.key + "_ADMIN", self.key + "_WRITE", self.key + "_READ"]
+
+        unadded_groups = []
+
+        for group_permission in zip(GROUP_NAMES, PERMISSIONS):
+            if group_permission[0] in groups:
+                group_project_permission = f"""curl -u'{self.user}' --request PUT --url 'https://{self.instance}/rest/api/latest/projects/{self.key}/permissions/groups?name={group_permission[0]}&permission={group_permission[1]}' """
+
+                try:
+                    self.run_post(group_project_permission)
+                except:
+                    unadded_groups.append(group_permission)
+            else:
+                unadded_groups.append(group_permission)
+
+        if len(unadded_groups) != 0:
+            print(f"These are the groups that could not be added to project {self.key}.")
+            for group_permission in unadded_groups:
+                print(group_permission)
+
+            print("\nCheck if project KEY is right. Or the groups exist")
 
 class Repository(Project):
     def __init__(self, instance, user, key, repo_name):
